+++ simulator/examples/image-tga.rs
@@ -1,6 +1,6 @@
 //! # Example: TGA images
 //!
-//! Draw a 16BPP TGA image to the display
+//! Draw a 24BPP TGA image to the display
 //!
 //! This example uses the [tinytga](https://crates.io/crates/tinytga) crate to load the TGA from a
 //! byte slice read in at compile time.
@@ -10,19 +10,18 @@
 
 use embedded_graphics::{image::Image, pixelcolor::Rgb888, prelude::*};
 use embedded_graphics_simulator::{OutputSettingsBuilder, SimulatorDisplay, Window};
-use tinytga::EgTga;
+use tinytga::Tga;
 
 fn main() -> Result<(), core::convert::Infallible> {
     let mut display: SimulatorDisplay<Rgb888> = SimulatorDisplay::new(Size::new(128, 128));
 
-    // Load the TGA image
-    let tga: EgTga<Rgb888> = EgTga::from_slice(include_bytes!("./assets/rust-pride.tga")).unwrap();
+    // Load the TGA image.
+    // The color type must be specified explicitly to match the color format used by the image,
+    // otherwise the compiler may infer an incorrect type.
+    let tga: Tga<Rgb888> = Tga::from_slice(include_bytes!("./assets/rust-pride.tga")).unwrap();
 
-    // TODO: update comment
-    // Create a new embedded-graphics Image, wrapping the TGA which provides pixel data. The top
-    // left corner of the image is positioned at (32, 32). It is important to specify the color
-    // format used by the image, otherwise the compiler may infer an incorrect type. This image is
-    // in 24BPP RGB888 format, so the Rgb888 pixel color type is used.
+    // To draw the `tga` object to the display it needs to be wrapped in an `Image` object to set
+    // the position at which it should drawn.
     let image = Image::new(&tga, Point::new(32, 32));
 
     // Display the image