+++ tinybmp/src/lib.rs
@@ -1,7 +1,7 @@
 //! A small BMP parser designed for embedded, no-std environments but usable anywhere. Beyond
 //! parsing the image header, no other allocations are made.
 //!
-//! To access the individual pixels in an image, the [`Bmp`] struct implements `IntoIterator`. It is
+//! To access the individual pixels in an image, the [`BmpRaw`] struct implements `IntoIterator`. It is
 //! also possible to access the raw image data by reading the [`pixel_data`] field.
 //!
 //! # Features
@@ -13,9 +13,9 @@
 //! ## Load a BMP image and check its [`Header`] and returned pixels.
 //!
 //! ```rust
-//! use tinybmp::{Bmp, FileType, Header, Pixel};
+//! use tinybmp::{BmpRaw, FileType, Header, Pixel};
 //!
-//! let bmp = Bmp::from_slice(include_bytes!("../tests/chessboard-8px-24bit.bmp"))
+//! let bmp = BmpRaw::from_slice(include_bytes!("../tests/chessboard-8px-24bit.bmp"))
 //!     .expect("Failed to parse BMP image");
 //!
 //! // Read the BMP header
@@ -53,13 +53,13 @@
 //! ```rust
 //! # #[cfg(feature = "graphics")] { fn main() -> Result<(), core::convert::Infallible> {
 //! use embedded_graphics::{image::Image, prelude::*};
-//! use tinybmp::EgBmp;
+//! use tinybmp::Bmp;
 //! # use embedded_graphics::mock_display::MockDisplay;
 //! # use embedded_graphics::pixelcolor::Rgb565;
 //! # let mut display: MockDisplay<Rgb565> = MockDisplay::default();
 //!
 //! // Load 16BPP 8x8px image
-//! let bmp: EgBmp<Rgb565> = EgBmp::from_slice(include_bytes!("../tests/chessboard-8px-color-16bit.bmp")).unwrap();
+//! let bmp: Bmp<Rgb565> = Bmp::from_slice(include_bytes!("../tests/chessboard-8px-color-16bit.bmp")).unwrap();
 //!
 //! let image = Image::new(&bmp, Point::zero());
 //!
@@ -69,8 +69,8 @@
 //!
 //! [embedded-graphics]: https://crates.io/crates/embedded-graphics
 //! [`Header`]: ./header/struct.Header.html
-//! [`Bmp`]: ./struct.Bmp.html
-//! [`pixel_data`]: ./struct.Bmp.html#structfield.pixel_data
+//! [`BmpRaw`]: ./struct.BmpRaw.html
+//! [`pixel_data`]: ./struct.BmpRaw.html#structfield.pixel_data
 
 #![no_std]
 #![deny(missing_docs)]
@@ -79,22 +79,28 @@
 mod header;
 mod pixel;
 
+#[cfg(feature = "graphics")]
+mod embedded_graphics;
+
 use crate::header::parse_header;
 pub use crate::{
     header::{FileType, Header},
     pixel::Pixel,
 };
 
+#[cfg(feature = "graphics")]
+pub use crate::embedded_graphics::*;
+
 /// A BMP-format bitmap
 #[derive(Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug)]
-pub struct Bmp<'a> {
+pub struct BmpRaw<'a> {
     /// Image header
     pub header: Header,
 
     image_data: &'a [u8],
 }
 
-impl<'a> Bmp<'a> {
+impl<'a> BmpRaw<'a> {
     /// Create a bitmap object from a byte array
     ///
     /// This method keeps a slice of the original input and does not dynamically allocate memory.
@@ -104,7 +110,7 @@ impl<'a> Bmp<'a> {
 
         let image_data = &bytes[header.image_data_start..];
 
-        Ok(Bmp { header, image_data })
+        Ok(BmpRaw { header, image_data })
     }
 
     /// Get a reference to the range of bytes that represents the pixel data in the image
@@ -142,7 +148,7 @@ impl<'a> Bmp<'a> {
     }
 }
 
-impl<'a> IntoIterator for &'a Bmp<'a> {
+impl<'a> IntoIterator for &'a BmpRaw<'a> {
     type Item = Pixel;
     type IntoIter = BmpIterator<'a>;
 
@@ -169,7 +175,7 @@ impl<'a> IntoIterator for &'a Bmp<'a> {
 #[derive(Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug)]
 pub struct BmpIterator<'a> {
     /// Reference to original BMP image
-    bmp: &'a Bmp<'a>,
+    bmp: &'a BmpRaw<'a>,
 
     /// Image pixel data as a byte slice, little endian ordering
     pixel_data: &'a [u8],
@@ -243,64 +249,3 @@ impl<'a> Iterator for BmpIterator<'a> {
         }
     }
 }
-
-#[cfg(feature = "graphics")]
-mod e_g {
-    use super::*;
-    use core::marker::PhantomData;
-    use embedded_graphics::{
-        pixelcolor::{raw::RawData, PixelColor},
-        prelude::*,
-    };
-
-    /// TODO: docs
-    #[derive(Debug)]
-    pub struct EgBmp<'a, C> {
-        bmp: Bmp<'a>,
-        color_type: PhantomData<C>,
-    }
-
-    impl<'a, C> EgBmp<'a, C>
-    where
-        C: PixelColor + From<<C as PixelColor>::Raw>,
-    {
-        /// TODO: docs
-        pub fn from_slice(data: &'a [u8]) -> Result<Self, ()> {
-            Ok(Self {
-                bmp: Bmp::from_slice(data)?,
-                color_type: PhantomData,
-            })
-        }
-    }
-
-    impl<C> ImageDrawable for EgBmp<'_, C>
-    where
-        C: PixelColor + From<<C as PixelColor>::Raw>,
-    {
-        type Color = C;
-
-        fn draw<D>(&self, target: &mut D) -> Result<(), D::Error>
-        where
-            D: DrawTarget<Color = C>,
-        {
-            target.fill_contiguous(
-                &self.bounding_box(),
-                self.bmp
-                    .into_iter()
-                    .map(|p| C::Raw::from_u32(p.color).into()),
-            )
-        }
-    }
-
-    impl<C> OriginDimensions for EgBmp<'_, C>
-    where
-        C: PixelColor,
-    {
-        fn size(&self) -> Size {
-            Size::new(self.bmp.width(), self.bmp.height())
-        }
-    }
-}
-
-#[cfg(feature = "graphics")]
-pub use e_g::*;