+++ tinytga/tests/embedded_graphics.rs
@@ -4,7 +4,7 @@ use embedded_graphics::{
     pixelcolor::{Gray8, Rgb888},
     prelude::*,
 };
-use tinytga::EgTga;
+use tinytga::Tga;
 
 const CHESSBOARD_PATTERN: &[&str] = &[
     "WKWK", //
@@ -15,7 +15,7 @@ const CHESSBOARD_PATTERN: &[&str] = &[
 
 #[test]
 fn chessboard_compressed() {
-    let tga: EgTga<Rgb888> = EgTga::from_slice(include_bytes!("./chessboard_4px_rle.tga")).unwrap();
+    let tga: Tga<Rgb888> = Tga::from_slice(include_bytes!("./chessboard_4px_rle.tga")).unwrap();
     let image = Image::new(&tga, Point::zero());
 
     let mut display = MockDisplay::new();
@@ -26,7 +26,7 @@ fn chessboard_compressed() {
 
 #[test]
 fn chessboard_uncompressed() {
-    let tga: EgTga<Rgb888> = EgTga::from_slice(include_bytes!("./chessboard_raw.tga")).unwrap();
+    let tga: Tga<Rgb888> = Tga::from_slice(include_bytes!("./chessboard_raw.tga")).unwrap();
     let image = Image::new(&tga, Point::zero());
 
     let mut display = MockDisplay::new();
@@ -36,7 +36,7 @@ fn chessboard_uncompressed() {
 }
 
 fn test_color_tga(data: &[u8]) {
-    let im: EgTga<Rgb888> = EgTga::from_slice(data).unwrap();
+    let im: Tga<Rgb888> = Tga::from_slice(data).unwrap();
     let image = Image::new(&im, Point::zero());
 
     let mut display = MockDisplay::new();
@@ -55,7 +55,7 @@ fn test_color_tga(data: &[u8]) {
 }
 
 fn test_gray_tga(data: &[u8]) {
-    let im: EgTga<Gray8> = EgTga::from_slice(data).unwrap();
+    let im: Tga<Gray8> = Tga::from_slice(data).unwrap();
     let image = Image::new(&im, Point::zero());
 
     let mut display = MockDisplay::new();